+++ CHANGES.md
@@ -3,7 +3,7 @@
 
 ### NEXT (Feature coming in next release, currently in master branch)
 - Add support for Python 3.9
-- Replace yaql in favour of jmespath for parsing query expressions
+- Replace yaql in favour of jsonpath-ng for parsing query expressions
 - Support nested query expressions in *args* and *kwargs* arguments to PythonTask (see example 11)
 
 ### 0.3.1